Add cache-busting query string
The HTML page with the list of PDFs is cached by CloudFlare for 1 hour,
and cache-control headers don't bypass it. Add a query parameter to bust
the cache. We already had situations where a new PDF is published and the
Lambda function fails to find it due to this cache.

We still do requests infrequently enough that this shouldn't cause extra
server load.