Deterministic build of the Lambda zip
Use strip-nondeterminism to remove timestamps and sort members in the zip
file. This makes the zip identical if files inside don't change, even if
their timestamps did change. For example, making a change, undoing it,
and building the zip will produce a bit-identical zip, and terraform
shouldn't re-upload it.

Note that strip-nondeterminism 1.1.2 doesn't work (some changing timestamp
is still kept) and 1.6.3 does. I don't know what's the exact version that
fixed the problem.